mavlink_dup: open backends within poll_state
it's possible that some backends are not (yet) connected. If so, open()
fails for these with ENOTCONN, but we still should let MavlinkDuplicator::
open() succeed, because other backends might be connected.

So open() is moved within poll_state, so that not yet connected backends
will work as well, once they're connected.

In my tests this was the case for ttyACM0, so USB serial.