mc_pos_control: keep landing gear down below 5 m
The landing gear takes some time to go down, so just listening for a
landing setpoint was not enough time to get it down in time, especially
when the last hover position is at e.g. 2.5 meters.
Therefore, this hardcodes the landing gear to be down whenever we are
closer than 5.0 meters to the home position. Obviously, this is not an
optimal solution, however it clears the problem for current flight
testing.